rename app.jsx -> App.jsx
